+++ revision-derive/src/expand/de.rs
@@ -99,8 +99,7 @@ impl<'ast> Visit<'ast> for DeserializeVisitor<'_> {
 			stream: &mut variants,
 			discriminants,
 		}
-		.visit_enum(i)
-		.unwrap();
+		.visit_enum(i)?;
 
 		let error_string =
 			format!("Invalid discriminant `{{x}}` for enum `{}` revision `{{__revision}}`", i.name);
@@ -131,8 +130,7 @@ impl<'ast> Visit<'ast> for DeserializeVisitor<'_> {
 			current: self.current,
 			stream: &mut fields_binding,
 		}
-		.visit_struct(i)
-		.unwrap();
+		.visit_struct(i)?;
 
 		match i.fields {
 			Fields::Named {
@@ -187,7 +185,12 @@ impl<'ast> Visit<'ast> for DeserializeVisitor<'_> {
 			f.attrs.options.exists_at(self.current) && !f.attrs.options.exists_at(self.target)
 		}) {
 			let binding = f.name.to_binding();
-			let convert = f.attrs.options.convert.as_ref().unwrap();
+			let convert = f
+				.attrs
+				.options
+				.convert
+				.as_ref()
+				.expect("FieldOptions::finish rejects `end` without convert_fn");
 			let convert = Ident::new(&convert.value(), convert.span());
 			let revision = self.current as u16;
 			self.stream.append_all(quote! {
@@ -223,8 +226,7 @@ impl<'ast> Visit<'ast> for DeserializeVariant<'_> {
 			current: self.current,
 			stream: &mut fields,
 		}
-		.visit_variant(i)
-		.unwrap();
+		.visit_variant(i)?;
 
 		let fields_struct_name = i.fields_name(&self.name.to_string());
 
@@ -257,7 +259,12 @@ impl<'ast> Visit<'ast> for DeserializeVariant<'_> {
 						&& !x.attrs.options.exists_at(self.target)
 				}) {
 					let binding = f.name.to_binding();
-					let convert = f.attrs.options.convert.as_ref().unwrap();
+					let convert = f
+						.attrs
+						.options
+						.convert
+						.as_ref()
+						.expect("FieldOptions::finish rejects `end` without convert_fn");
 					let convert = Ident::new(&convert.value(), convert.span());
 					let revision = self.current as u16;
 					bindings.append_all(quote! {
@@ -299,7 +306,12 @@ impl<'ast> Visit<'ast> for DeserializeVariant<'_> {
 						&& !x.attrs.options.exists_at(self.target)
 				}) {
 					let binding = f.name.to_binding();
-					let convert = f.attrs.options.convert.as_ref().unwrap();
+					let convert = f
+						.attrs
+						.options
+						.convert
+						.as_ref()
+						.expect("FieldOptions::finish rejects `end` without convert_fn");
 					let convert = Ident::new(&convert.value(), convert.span());
 					let revision = self.current as u16;
 					bindings.append_all(quote! {
@@ -339,7 +351,12 @@ impl<'ast> Visit<'ast> for DeserializeVariant<'_> {
 				.discriminants
 				.get(&i.ident)
 				.expect("missed variant during discriminant calculation");
-			let convert = i.attrs.options.convert.as_ref().unwrap();
+			let convert = i
+				.attrs
+				.options
+				.convert
+				.as_ref()
+				.expect("VariantOptions::finish rejects `end` without convert_fn");
 			let convert = Ident::new(&convert.value(), convert.span());
 			let revision = self.current as u16;